add mfcc to dataset and reorganize supporting code
